feat: add fileSearch tool
