+++ packages/ai-native/src/browser/index.ts
@@ -58,7 +58,7 @@ import { MCPServerProxyService } from './mcp/mcp-server-proxy.service';
 import { MCPServerRegistry } from './mcp/mcp-server.feature.registry';
 import { CreateNewFileWithTextTool } from './mcp/tools/createNewFileWithText';
 import { EditFileTool } from './mcp/tools/editFile';
-import { FindFilesByNameSubstringTool } from './mcp/tools/findFilesByNameSubstring';
+import { FileSearchTool } from './mcp/tools/fileSearch';
 import { GetCurrentFilePathTool } from './mcp/tools/getCurrentFilePath';
 import { GetDiagnosticsByPathTool } from './mcp/tools/getDiagnosticsByPath';
 import { GetFileTextByPathTool } from './mcp/tools/getFileTextByPath';
@@ -105,7 +105,7 @@ export class AINativeModule extends BrowserModule {
     GetOpenEditorFileTextTool,
     GetFileTextByPathTool,
     GetCurrentFilePathTool,
-    FindFilesByNameSubstringTool,
+    FileSearchTool,
     GetDiagnosticsByPathTool,
     RunTerminalCommandTool,
     ReplaceOpenEditorFileByDiffPreviewerTool,
